kovan: Move to newer CPU ID
The CPU ID, as returned by the read_cpu_id() function in Linux,
returns a vendor-specific flag indicating which ID we're using.

This patch removes the old S0-rev ID and moves towards the newer
A0- and B0- CPU ID.  As such, it requires both a kernel config
patch and a U-Boot patch.

This enables us to use mainline kernels with significantly fewer
changes to various critical components.