mysql5: Fix startup conflict for network target
Previously, mysql5-server was using a SysV-style init script.  This was
causing a circular dependency for systemd, which was solving the loop by
removing all of the networking services.

We get around this by adding a service file for mysqld, which was largely
copied from Fedora.  This breaks the circular dependency, allowing the
system to boot normally.