added imx6q.dtsi change for microSD CD bypass. Note ugly buildbot integration.
I'm following on Sean's pattern to copy files into a certain branch
of a linux kernel, instead of having a checked out branch of our own
which I think is the right thing to do. See, the problem is that we're
just copying source files over a bit at a time and that just "feels wrong".